Pasar grafica a fragmento
