+++ .idea/modules.xml
@@ -4,7 +4,6 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/ProyectoMoviles.iml" filepath="$PROJECT_DIR$/ProyectoMoviles.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/lifetqueremos.iml" filepath="$PROJECT_DIR$/lifetqueremos.iml" />
       <module fileurl="file://C:\Users\Android\Documents\TareasTec\Semestre6\Android\ProyectoMoviles\lifetqueremos.iml" filepath="C:\Users\Android\Documents\TareasTec\Semestre6\Android\ProyectoMoviles\lifetqueremos.iml" />
     </modules>
   </component>
